split out uploads and downloads controllers
