+++ config/routes.rb
@@ -1,4 +1,4 @@
 Rails.application.routes.draw do
-  get '/service/:service_slug/user/:user_id/:fingerprint_with_prefix', to: 'user_file#show'
-  post '/service/:service_slug/user/:user_id', to: 'user_file#create'
+  get '/service/:service_slug/user/:user_id/:fingerprint_with_prefix', to: 'downloads#show'
+  post '/service/:service_slug/user/:user_id', to: 'uploads#create'
 end
